added some funtions for computer/bot
added the methods that runGame method that takes in what the player choice is as a string. and rolls a 3 sided die to see what the cumpooter will pick randomly. after teh computer makes a choice the method also has if statements to see what the apropriate response will be.